+++ R/digital-ocean.R
@@ -31,9 +31,17 @@
 #'  - A 4GB swap file is created to ensure that machines with little RAM (the default) are
 #'    able to get through the necessary R package compilations.
 #' @export
+#' @examples \dontrun{
+#' droplet = do_provision()
+#' analogsea::droplets()
+#' analogsea::install_r_package(droplet, c("readr", "remotes"))
+#' do_deploy_api(droplet, "hello",
+#' system.file("plumber", "10-welcome", package = "plumber"),
+#' port=8000, forward=TRUE)
+#' }
 do_provision <- function(droplet, unstable=FALSE, example=TRUE, ...){
 
-  if (missing(droplet)){
+  if (missing(droplet) || is.null(droplet)){
 
     # No droplet provided; create a new server
     message("THIS ACTION COSTS YOU MONEY!")
@@ -293,9 +301,19 @@ do_configure_https <- function(droplet, domain, email, termsOfService=FALSE, for
 #' @param ... additional arguments to pass to [analogsea::droplet_ssh()] or
 #'   [analogsea::droplet_upload()], such as `keyfile`.
 #'   Cannot contain `remote`, `local` as named arguments.
+#' @param overwrite if an application is already running for this `path` name,
+#'   and `overwrite = TRUE`, then `do_remove_api` will be run.
 #' @export
 do_deploy_api <- function(droplet, path, localPath, port, forward=FALSE,
-                          docs=FALSE, preflight, ...){
+                          docs=FALSE, preflight, overwrite = FALSE, ...){
+  args = list(...)
+  if ("swagger" %in% names(args)) {
+    lifecycle::deprecate_stop("0.1.3",
+                              "do_deploy_api(swagger = )",
+                              "do_deploy_api(docs = )")
+  }
+
+
   # Trim off any leading slashes
   path <- sub("^/+", "", path)
   # Trim off any trailing slashes if any exist.
@@ -315,6 +333,34 @@ do_deploy_api <- function(droplet, path, localPath, port, forward=FALSE,
   remoteTmp <- paste0("/tmp/",
                       paste0(sample(LETTERS, 10, replace=TRUE), collapse=""))
   dirName <- basename(localPath)
+
+  plumber_path = paste0("/var/plumber/", path)
+
+  # removing the path in case it already exists (may want to error here)
+  # analogsea::droplet_ssh(droplet, paste0("rm -rf ", plumber_path), ...)
+  if (overwrite) {
+    output = try({
+      do_remove_api(droplet, path = path, delete = TRUE, ...)
+    }, silent = TRUE)
+    if (inherits(output, "try-error")) {
+      msg = paste0("Tried to remove ", path, " application had issues")
+      warning(msg)
+    }
+  }
+
+  cmd = paste0("if [ -f ", plumber_path, " ]; then echo 'TRUE'; else echo 'FALSE'; fi")
+  check_path =
+    utils::capture.output({
+      analogsea::droplet_ssh(droplet, cmd, ...)
+    }, type = "output")
+  check_path = check_path[ check_path %in% c("TRUE", "FALSE")]
+  check_path = as.logical(check_path)
+  if (check_path) {
+      stop(paste0(plumber_path, " already exists, either rename ",
+                  "or run do_remove_api"))
+  }
+
+
   analogsea::droplet_ssh(droplet, paste0("mkdir -p ", remoteTmp), ...)
   analogsea::droplet_upload(droplet, local=localPath, remote=remoteTmp, ...)
   analogsea::droplet_ssh(droplet,
@@ -428,8 +474,11 @@ do_forward <- function(droplet, path){
 #' @param path The path/name of the plumber service
 #' @param delete If `TRUE`, will also delete the associated directory
 #'   (`/var/plumber/whatever`) from the server.
+#' @param ... additional arguments to pass to [analogsea::droplet_ssh()] or
+#'   [analogsea::droplet_upload()], such as `keyfile`.
+#'   Cannot contain `remote`, `local` as named arguments.
 #' @export
-do_remove_api <- function(droplet, path, delete=FALSE){
+do_remove_api <- function(droplet, path, delete=FALSE, ...){
   # Trim off any leading slashes
   path <- sub("^/+", "", path)
   # Trim off any trailing slashes if any exist.
@@ -446,18 +495,27 @@ do_remove_api <- function(droplet, path, delete=FALSE){
   if (nchar(path)==0){
     stop("Path cannot be empty.")
   }
-
+  try_ssh = function(...) {
+    try(analogsea::droplet_ssh(...), silent = TRUE)
+  }
   serviceName <- paste0("plumber-", path)
-  analogsea::droplet_ssh(droplet, paste0("systemctl stop ", serviceName))
-  analogsea::droplet_ssh(droplet, paste0("systemctl disable ", serviceName))
-  analogsea::droplet_ssh(droplet, paste0("rm /etc/systemd/system/", serviceName, ".service"))
-  analogsea::droplet_ssh(droplet, paste0("rm /etc/nginx/sites-available/plumber-apis/", path, ".conf"))
+  message("stopping service: ", serviceName)
+  try_ssh(droplet, paste0("systemctl stop ", serviceName), ...)
+  message("disabling service: ", serviceName)
+  try_ssh(droplet, paste0("systemctl disable ", serviceName), ...)
+  message("removing service: ", serviceName)
+  try_ssh(droplet, paste0("rm /etc/systemd/system/", serviceName, ".service"), ...)
+  message("removing config: ", serviceName)
+  try_ssh(droplet, paste0("rm /etc/nginx/sites-available/plumber-apis/", path, ".conf"), ...)
 
-  analogsea::droplet_ssh(droplet, "systemctl reload nginx")
+  message("reloading nginx: ", serviceName)
+  try_ssh(droplet, "systemctl reload nginx", ...)
 
   if(delete){
-    analogsea::droplet_ssh(droplet, paste0("rm -rf /var/plumber/", path))
+    message("reloading plumber folder: ", serviceName)
+    try_ssh(droplet, paste0("rm -rf /var/plumber/", path), ...)
   }
+  return(droplet)
 }
 
 #' Remove the forwarding rule